requirements.txt
# Data and training
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
easyocr>=1.7.0
Pillow>=10.0.0
opencv-python>=4.8.0
pillow-heif>=0.18.0
